fixed int double weird numbers bug. also fixed blank coin bug now all storms send with proper first coin (were converting to int before) when fetching coins they will now all be properly formatted into two decimal places

+++ storm/DBManager.h
@@ -15,7 +15,7 @@
 
 + (void) authenticate;
 
-+ (void)createStorm:(NSString*)toUser withMessage:(NSString*)message withCoinValue:(int)coinValue toRecipient:(Friend*)recipient;
++ (void)createStorm:(NSString*)toUser withMessage:(NSString*)message withCoinValue:(double)coinValue toRecipient:(Friend*)recipient;
 + (void) sendCoin:(double)coinValue withStorm:(Storm*)storm;
 
 @end
